rename: 역할에 맞게 Mapper 클래스로 이름 변경
- GroupUtil에서 GroupResponseMapper 클래스로 이름을 변경
- SchoolUtil에서 SchoolResponseMapper 클래스로 이름을 변경
